Added warnings if the LLM returns something with the wrong JSON structure, and moved the Classifier class out to its own file
